automod: refactor capture (also fetch-and-process) out of engine pkg
