+++ automod/capture/capture.go
@@ -1,22 +1,21 @@
-package engine
+package capture
 
 import (
 	"context"
 
 	comatproto "github.com/bluesky-social/indigo/api/atproto"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/bluesky-social/indigo/automod"
 )
-
-// REVIEW: if this "capture" code can leave the engine package.  It seems likely.
 
 type AccountCapture struct {
 	CapturedAt  syntax.Datetime                     `json:"capturedAt"`
-	AccountMeta AccountMeta                         `json:"accountMeta"`
+	AccountMeta automod.AccountMeta                 `json:"accountMeta"`
 	PostRecords []comatproto.RepoListRecords_Record `json:"postRecords"`
 }
 
-func (e *Engine) CaptureRecent(ctx context.Context, atid syntax.AtIdentifier, limit int) (*AccountCapture, error) {
-	ident, records, err := e.FetchRecent(ctx, atid, limit)
+func CaptureRecent(ctx context.Context, eng *automod.Engine, atid syntax.AtIdentifier, limit int) (*AccountCapture, error) {
+	ident, records, err := FetchRecent(ctx, eng, atid, limit)
 	if err != nil {
 		return nil, err
 	}
@@ -29,7 +28,7 @@ func (e *Engine) CaptureRecent(ctx context.Context, atid syntax.AtIdentifier, li
 
 	// clear any pre-parsed key, which would fail to marshal as JSON
 	ident.ParsedPublicKey = nil
-	am, err := e.GetAccountMeta(ctx, ident)
+	am, err := eng.GetAccountMeta(ctx, ident)
 	if err != nil {
 		return nil, err
 	}
