+++ automod/capture/capture_test.go
@@ -1,16 +1,17 @@
-package engine
+package capture
 
 import (
 	"testing"
 
 	"github.com/bluesky-social/indigo/automod/countstore"
+	"github.com/bluesky-social/indigo/automod/engine"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestNoOpCaptureReplyRule(t *testing.T) {
 	assert := assert.New(t)
 
-	eng := EngineTestFixture()
+	eng := engine.EngineTestFixture()
 	capture := MustLoadCapture("testdata/capture_atprotocom.json")
 	assert.NoError(ProcessCaptureRules(&eng, capture))
 	c, err := eng.GetCount("automod-quota", "report", countstore.PeriodDay)
