+++ automod/capture/fetch.go
@@ -1,4 +1,4 @@
-package engine
+package capture
 
 import (
 	"context"
@@ -7,15 +7,16 @@ import (
 	comatproto "github.com/bluesky-social/indigo/api/atproto"
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/bluesky-social/indigo/automod"
 	"github.com/bluesky-social/indigo/xrpc"
 )
 
-func (e *Engine) FetchAndProcessRecord(ctx context.Context, aturi syntax.ATURI) error {
+func FetchAndProcessRecord(ctx context.Context, eng *automod.Engine, aturi syntax.ATURI) error {
 	// resolve URI, identity, and record
 	if aturi.RecordKey() == "" {
 		return fmt.Errorf("need a full, not partial, AT-URI: %s", aturi)
 	}
-	ident, err := e.Directory.Lookup(ctx, aturi.Authority())
+	ident, err := eng.Directory.Lookup(ctx, aturi.Authority())
 	if err != nil {
 		return fmt.Errorf("resolving AT-URI authority: %v", err)
 	}
@@ -25,7 +26,7 @@ func (e *Engine) FetchAndProcessRecord(ctx context.Context, aturi syntax.ATURI) 
 	}
 	pdsClient := xrpc.Client{Host: ident.PDSEndpoint()}
 
-	e.Logger.Info("fetching record", "did", ident.DID.String(), "collection", aturi.Collection().String(), "rkey", aturi.RecordKey().String())
+	eng.Logger.Info("fetching record", "did", ident.DID.String(), "collection", aturi.Collection().String(), "rkey", aturi.RecordKey().String())
 	out, err := comatproto.RepoGetRecord(ctx, &pdsClient, "", aturi.Collection().String(), ident.DID.String(), aturi.RecordKey().String())
 	if err != nil {
 		return fmt.Errorf("fetching record from Relay (%s): %v", aturi, err)
@@ -34,19 +35,19 @@ func (e *Engine) FetchAndProcessRecord(ctx context.Context, aturi syntax.ATURI) 
 		return fmt.Errorf("expected a CID in getRecord response")
 	}
 	recCID := syntax.CID(*out.Cid)
-	op := RecordOp{
-		Action:     CreateOp,
+	op := automod.RecordOp{
+		Action:     automod.CreateOp,
 		DID:        ident.DID,
 		Collection: aturi.Collection(),
 		RecordKey:  aturi.RecordKey(),
 		CID:        &recCID,
 		Value:      out.Value.Val,
 	}
-	return e.ProcessRecordOp(ctx, op)
+	return eng.ProcessRecordOp(ctx, op)
 }
 
-func (e *Engine) FetchRecent(ctx context.Context, atid syntax.AtIdentifier, limit int) (*identity.Identity, []*comatproto.RepoListRecords_Record, error) {
-	ident, err := e.Directory.Lookup(ctx, atid)
+func FetchRecent(ctx context.Context, eng *automod.Engine, atid syntax.AtIdentifier, limit int) (*identity.Identity, []*comatproto.RepoListRecords_Record, error) {
+	ident, err := eng.Directory.Lookup(ctx, atid)
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to resolve AT identifier: %v", err)
 	}
@@ -60,13 +61,13 @@ func (e *Engine) FetchRecent(ctx context.Context, atid syntax.AtIdentifier, limi
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to fetch record list: %v", err)
 	}
-	e.Logger.Info("got recent posts", "did", ident.DID.String(), "pds", pdsURL, "count", len(resp.Records))
+	eng.Logger.Info("got recent posts", "did", ident.DID.String(), "pds", pdsURL, "count", len(resp.Records))
 	return ident, resp.Records, nil
 }
 
-func (e *Engine) FetchAndProcessRecent(ctx context.Context, atid syntax.AtIdentifier, limit int) error {
+func FetchAndProcessRecent(ctx context.Context, eng *automod.Engine, atid syntax.AtIdentifier, limit int) error {
 
-	ident, records, err := e.FetchRecent(ctx, atid, limit)
+	ident, records, err := FetchRecent(ctx, eng, atid, limit)
 	if err != nil {
 		return err
 	}
@@ -78,15 +79,15 @@ func (e *Engine) FetchAndProcessRecent(ctx context.Context, atid syntax.AtIdenti
 			return fmt.Errorf("parsing PDS record response: %v", err)
 		}
 		recCID := syntax.CID(rec.Cid)
-		op := RecordOp{
-			Action:     CreateOp,
+		op := automod.RecordOp{
+			Action:     automod.CreateOp,
 			DID:        ident.DID,
 			Collection: aturi.Collection(),
 			RecordKey:  aturi.RecordKey(),
 			CID:        &recCID,
 			Value:      rec.Value.Val,
 		}
-		err = e.ProcessRecordOp(ctx, op)
+		err = eng.ProcessRecordOp(ctx, op)
 		if err != nil {
 			return err
 		}
